routing: introduce routing declaration language
This is a simple micro-language specifically designed for describing URL
routes. Before this, URLconf had to be declared in JSON, which is barely
maintainable and error-prone.

The syntax is Python-inspired and features significant indentation. The
indentation rule is intentionally made to follow Python's, i.e. tab stop
is fixed at 8. Route declaration is of the following format:

    --<type of router, e.g. regex or exact>
    <target list>

where "target list" consists of lines or blocks of individual route
targets and sub-router declarations. A simple target is defined as

    <pattern> <endpoint> <renderer> [extra info]

while a sub-router declaration is like

    <pattern> [extra info]:
    	<route declaration>

note that the declaration of sub-router is indented. Items surrounded by
brackets are optional; they default to None. A value of "null"
(case-sensitive) in the "renderer" field makes it disappear in the final
extra data field.

The strings can be quoted to allow whitespaces inside, but no serious
escaping is implemented; this can be a problem if you actually want to
embed quotation marks into URL. This limitation may be lifted in a
future revision, or, depending on need, may not.