Add '--include=xxx' directive for urlconf
This adds to weiyu the ability to split one urlconf file into many.
Usage is to put ONLY this directive inside a sub-router definition, like
this:

/something/:
    --include=urls.something.txt

This will load urls.something.txt in the current working directory and
treat its content as definitions for another root-level router (thus
attributes like default renderer are NOT passed through). Both the file
path and inheritance limitations are expected to be fixed.