utils: Add view loader
This is an auto-importer for modules, designed to move all the import
statements out of your adapter file. It can be configured in JSON, and
can do either absolute imports or relative imports based on a package
specified.

The modules specified in the `modules` key are loaded in order,
optionally with the package specified by `package` as anchor for
relative import. If `package` key is absent, or converts to `None`, the
module would only do absolute imports (this is exactly the same behavior
the underlying `importlib.import_module` exhibits).