shortcuts: initialization made more convenient
For developers' convenience, the adapter stub cannot be simplified
enough. So I decided to move all those "naming conventions" and
initialization boilerplate to the shortcuts package. To further reduce
the number of keystrokes required (and amount of injury done to the
fingers), the application object is also injected using `inspect`
functionality.