weiyu: remove old and broken auth package
The feature was on the original specification for weiyu/luohua project.
However a year later I realized the originally conceived way of doing it
in the framework is plain wrong; the whole thing should be ideally
managed from the app-side. The proper abstraction is yet to come up, so
the package is removed for the moment.